+++ shop/models.py
@@ -44,7 +44,7 @@ class Product(models.Model):
     code = models.UUIDField(primary_key=True,default=uuid.uuid4, editable=False)
     image = models.ImageField(upload_to='products/%Y/%m/%d')
     description = models.TextField()
-    price = models.DecimalField(max_digits=10,decimal_places=3)
+    price = models.IntegerField()
     available = models.BooleanField(default=True)
     created = models.DateTimeField(auto_now_add=True)
     updated = models.DateTimeField(auto_now=True)
